fix(widget-cald): arithmetic do empacotamento em CGFloat
Erro real de build: `columnSpacing * (col.isEmpty ? 0 : col.count - 1)` —
CGFloat multiplicado por Int a partir do nº de gaps. Corrige convertendo a
contagem p/ CGFloat na hora. Constantes de altura já eram CGFloat; somas de
altura acumulada e a altura útil (systemMedium - 26*2) já eram CGFloat de
ponta a ponta.

+++ ios/WidgetExtension/WidgetExtension.swift
@@ -879,7 +879,9 @@ struct CalendarLayout {
     // com o mesmo espacamento do empacotamento — usada na verificação DEBUG.
     static func renderedColumnHeight(_ col: [CalendarSlotView]) -> CGFloat {
         let sum = col.reduce(0) { $0 + slotHeight($1) }
-        return sum + columnSpacing * (col.isEmpty ? 0 : col.count - 1)
+        // Nº de gaps entre células (contagem), convertido na hora p/ o * —
+        // só este operando é inteiro por natureza (conta posições).
+        return sum + columnSpacing * CGFloat(col.isEmpty ? 0 : col.count - 1)
     }
 
     private static func slotHeight(_ slot: CalendarSlotView) -> CGFloat {
